make logger browser compatible
